requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
parse==1.19.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.2.1
tqdm==4.60.0
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
